backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment Variables
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Email (Optional)
aiosmtplib==3.0.1

# AI/ML (Optional - for future integration)
scikit-learn==1.4.0
pandas==2.2.0

# Utilities
python-dateutil==2.8.2